pyproject.toml
[project]
name = "whisper-save"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "orjson>=3.11.3",
    "pycryptodome>=3.23.0",
]

[tool.uv]
package = true
